% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_stereo.R
\docType{data}
\name{coord_azimuth_zenith}
\alias{coord_azimuth_zenith}
\alias{CoordAzimuthZenith}
\title{An azimuth/zenith coordinate system}
\format{An object of class \code{CoordAzimuthZenith} (inherits from \code{Coord}, \code{ggproto}, \code{gg}) of length 16.}
\usage{
coord_azimuth_zenith(projection = c("stereographic", "orthographic"))

CoordAzimuthZenith
}
\arguments{
\item{projection}{Use stereographic to look at the bottom of the sphere as if you are
situated at the top of the sphere (probably what you want); use orthographic to look
at the bottom of the sphere as if you were very far away from it.}
}
\description{
This coordinate system imagines that your data will be plotted on the bottom of a
sphere, oriented such that 0 azimuth is at the positive x-axis, and 0 zenith is at the
top of the bowl. Increasing azimuth is counterclockwise, and zenith angles that are more
negative move to the center of the plot. Positive zenith angles are not relevant in this
coordinate system, and are squished to zero with a warning. A more approachable
coordinate system is \link{coord_bearing_dip}.
}
\examples{
library(ggplot2)

df <- data.frame(
  azimuth = seq(0, 4*pi, length.out = 40),
  zenith = seq(0, -pi / 2, length.out = 40)
)

ggplot(df, aes(azimuth, zenith)) +
  geom_path(col = "red") +
  geom_point(col = "blue") +
  coord_azimuth_zenith() +
  scale_x_continuous(
    breaks = radians(seq(0, 330, by = 30)),
    labels = degrees
  )


}
\keyword{datasets}
