% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{coord_bearing_plunge}
\alias{coord_bearing_plunge}
\title{A bearing/dip coordinate system}
\usage{
coord_bearing_plunge(projection = c("stereographic", "orthographic"))
}
\arguments{
\item{projection}{Use stereographic to look at the bottom of the sphere as if you are
situated at the top of the sphere (probably what you want); use orthographic to look
at the bottom of the sphere as if you were very far away from it.}
}
\description{
This coordinate system imagines that your data will be plotted on the bottom of a
sphere, as viewed from the inside, oriented such that 0 bearing is due north
and and 0 plunge is horizontal. Increasing plunge values plot towards the centre of
the plot, and increasing bearing values plot clockwise from north.
}
\examples{
library(ggplot2)

df <- data.frame(
  bearing = seq(0, 4*180, length.out = 40),
  plunge = seq(0, 90, length.out = 40)
)

ggplot(df, aes(bearing, plunge)) +
  geom_path(col = "red") +
  geom_point(col = "blue") +
  coord_bearing_plunge() +
  scale_x_continuous(breaks = seq(0, 330, by = 30))

}
