% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deep_hollow}
\alias{deep_hollow}
\title{Bedding and cleavage measurements from Deep Hollow, Kings County, Nova Scotia}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66 rows and 4 columns.}
\source{
White, C.E. (2010). Pre-Carboniferous Bedrock Geology of the
Annapolis Valley Area (NTS 21A/14, 15 and 16; 21H/01 and 02), Southern Nova
Scotia. In Mineral Resources Branch, Report of Activities 2009; Nova Scotia
Department of Natural Resources, Report ME 2010-1, p. 137-155.
\url{https://novascotia.ca/natr/meb/data/pubs/10re01/10re01_18White.pdf} in
\url{https://novascotia.ca/natr/meb/pdf/10re01.asp}
}
\usage{
deep_hollow
}
\description{
A sample dataset of strike and dip measurements (in degrees) for bedding and
cleavage planes in the Deep Hollow area, Kings County, Nova Scotia.
}
\keyword{datasets}
