% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_plane.R
\docType{data}
\name{geom_plane}
\alias{geom_plane}
\alias{GeomPlane}
\title{Draw a plane on a steronet}
\format{An object of class \code{GeomPlane} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
geom_plane(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., n = 101, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)

GeomPlane
}
\arguments{
\item{mapping}{A mapping}

\item{data}{Data}

\item{position}{Which position to use}

\item{...}{Set aesthetics}

\item{n}{Number of points to use to generate path}

\item{na.rm}{Remove points with missing values}

\item{show.legend}{Show legend}

\item{inherit.aes}{Inherit aesthetics from the plot}

\item{geom}{Which geometry to use}
}
\description{
Draw a plane on a steronet
}
\keyword{datasets}
