% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{radians}
\alias{radians}
\alias{degrees}
\title{Convert to/from degrees and radians}
\usage{
radians(degrees)

degrees(radians)
}
\arguments{
\item{degrees}{A value in degrees}

\item{radians}{A value in radians}
}
\description{
Convert to/from degrees and radians
}
\examples{
radians(180)
degrees(pi)

}
