% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_stereo.R
\name{scale_x_azimuth}
\alias{scale_x_azimuth}
\alias{pi_labels}
\title{Pretty breaks for coord_azimuth_zenith}
\usage{
scale_x_azimuth(breaks = seq(0, 7 * pi/4, by = pi/4),
  labels = pi_labels, ...)

pi_labels(breaks, ...)
}
\arguments{
\item{breaks}{Where break lines and labels are drawn}

\item{labels}{Label function that produces labels in terms of pi}

\item{...}{Passed to \link[ggplot2]{scale_x_continuous}.}
}
\description{
Pretty breaks for coord_azimuth_zenith
}
