% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_stereo.R
\name{scale_x_bearing}
\alias{scale_x_bearing}
\alias{scale_x_compass}
\title{Pretty breaks for coord_bearing_plunge}
\usage{
scale_x_bearing(breaks = seq(0, 330, by = 30), minor_breaks = seq(0,
  350, by = 10), ...)

scale_x_compass(breaks = c(0, 90, 180, 270), labels = c("N", "E", "S",
  "W"), minor_breaks = seq(0, 350, by = 10), ...)
}
\arguments{
\item{breaks, minor_breaks}{Where break lines and labels are drawn}

\item{...}{Passed to \link[ggplot2]{scale_x_continuous}.}
}
\description{
Pretty breaks for coord_bearing_plunge
}
