% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deep_hollow}
\alias{deep_hollow}
\title{Bedding and cleavage measurements from Deep Hollow, Kings County, Nova Scotia}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66 rows and 4 columns.}
\source{
\url{https://novascotia.ca/natr/meb/data/pubs/10re01/10re01_18White.pdf}
}
\usage{
deep_hollow
}
\description{
A sample dataset of strike and dip measurements (in degrees) for bedding and
cleavage planes in the Deep Hollow area, Kings County, Nova Scotia.
}
\keyword{datasets}
