% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{scale_x_azimuth}
\alias{scale_x_azimuth}
\alias{pi_labels}
\title{Pretty breaks for coord_azimuth_zenith}
\usage{
scale_x_azimuth(breaks = seq(0, 7 * pi/4, by = pi/4),
  labels = pi_labels, ...)

pi_labels(breaks, ...)
}
\arguments{
\item{breaks}{Where break lines and labels are drawn}

\item{labels}{Labels or label function that produces labels from breaks}

\item{...}{Passed to \link[ggplot2]{scale_x_continuous}.}
}
\description{
Pretty breaks for coord_azimuth_zenith
}
